Remove the event-id hash in hyperlith
It was used to prevent sending the repsonse if the content on the
client was exactly the same as on the server. However, I now think it's an unnecessary optimisation that adds overhead (to the regular usecase) and doesn't really save much. The compression is good enough that it's not really saving bandwidth, the transfer and compression isn't that expensive compared to generating the templates/querying the db and it assumes you haven't made any client side changes.