Add gridless example of virtual scroll
This version of checkboxes uses relative positioning
and translate instead of a large grid.

Grids are only used for the "9 patch" view and
the 9 chunks that compose it.

The advantage of this version is you can use more fine
grained chunks on the view and it lets you scale past max
grid size rendering limits. The grid approach falls
down around a billion grid cells as the browser can't
handle it.

With this approach there is no theoretical limit to the maximum
number of chunks the dataset can be.

The reason for using translate is it's a lot more
performant than top/left as it uses the GPU. It also
significantly reduces layout shifts.